requirements.txt
glcontext==2.3.5
imageio==2.15.0
imageio-ffmpeg==0.4.7
moderngl==5.6.4
moderngl-window==2.4.1
mttkinter==0.6.1
multipledispatch==0.6.0
numpy==1.19.5
Pillow==8.4.0
pkg_resources==0.0.0
pyglet==1.5.23
pyrr==0.10.3
pyzmq==22.3.0
six==1.16.0
tkVideo==0.1
zmq==0.0.0
